pyproject.toml
[project]
name = "ai-sidekick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.6.3",
    "openai>=1.99.1",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "wikipedia-api>=0.8.1",
]
